Move map into UI
