+++ public/game/map.js
@@ -50,7 +50,7 @@ function randomEncounter() {
 // Look for a free node in the next row to the right of the "desired index".
 const isEncounter = (node) => node && Boolean(node.type)
 
-function findPath(graph, graphEl, desiredIndex) {
+export function findPath(graph, graphEl, desiredIndex) {
 	let path = []
 
 	console.groupCollapsed('drawing path', desiredIndex)
@@ -125,7 +125,7 @@ function findPath(graph, graphEl, desiredIndex) {
 	return path
 }
 
-function drawPath(graph, path, graphEl) {
+export function drawPath(graph, path, graphEl) {
 	const svg = graphEl.querySelector('svg.paths')
 	console.group('drawPath')
 	path.forEach((move, index) => {
@@ -170,83 +170,6 @@ function getPosWithin(el, container) {
 		height: rect.height,
 	}
 }
-
-// This is an example of how you can render the graph as a map.
-class SlayMap extends HTMLElement {
-	connectedCallback() {
-		this.state = {
-			rows: this.getAttribute('rows'),
-			columns: this.getAttribute('columns'),
-		}
-		this.render()
-	}
-	render() {
-		const graph = generateGraph(this.state.rows, this.state.columns)
-		this.innerHTML = `
-			${graph
-				.map(
-					(row) => `
-				<slay-map-row>
-					${row
-						.map((col) => {
-							if (col.type === 'start') {
-								return `<slay-map-encounter>start</slay-map-encounter>`
-							}
-							if (col.type === 'end') {
-								return `<slay-map-encounter>end</slay-map-encounter>`
-							}
-							if (col.type) {
-								return `<slay-map-encounter>${col.type}</slay-map-encounter>`
-							}
-							return `<slay-map-node></slay-map-node>`
-						})
-						.join('')}
-				</slay-map-row>
-			`
-				)
-				.join('')}
-			<svg class="paths"></svg>
-		`
-		if (!graph[0][0].el) this.addElementsToGraph(graph)
-		this.scatter()
-		this.drawPaths(graph)
-		console.log({graph})
-	}
-	addElementsToGraph(graph) {
-		graph.forEach((row, rowIndex) => {
-			row
-				.filter((n) => n.type)
-				.forEach((node, nodeIndex) => {
-					// nth-of starts at 1, not 0
-					node.el = this.querySelector(
-						`slay-map-row:nth-of-type(${rowIndex + 1})
-							 slay-map-encounter:nth-of-type(${nodeIndex + 1})`
-					)
-				})
-		})
-	}
-	// Move the encounters around a bit.
-	scatter() {
-		const nodes = this.querySelectorAll('slay-map-encounter')
-		nodes.forEach((node) => {
-			node.style.transform = `translate3d(
-				${randomBetween(-35, 35)}%,
-				${randomBetween(-35, 35)}%,
-				0)
-			`
-		})
-	}
-	drawPaths(graph) {
-		// around ~90-140ms
-		console.time('mapRender')
-		drawPath(graph, findPath(graph, this, 0), this)
-		// drawPath(graph, findPath(graph, this, 2), this)
-		// drawPath(graph, findPath(graph, this, 3), this)
-		drawPath(graph, findPath(graph, this, 5), this)
-		console.timeEnd('mapRender')
-	}
-}
-customElements.define('slay-map', SlayMap)
 
 // https://github.com/oskarrough/slaytheweb/issues/28
 // https://i.imgur.com/oAofMa0.jpg
