+++ public/index.html
@@ -21,6 +21,7 @@
 			<slay-map rows="10" columns="6"></slay-map>
 		</div>
 		<script async src="game/map.js" type="module"></script>
+		<script async src="ui/map.js" type="module"></script>
 		<script>
 			// Tool to regenerate the map for testing.
 			if (!window.stw) window.stw = {}
